Move VAD enumeration into pe_symbols
